Enable projects that have valid targets anywhere in the csproj
Previously, the final import in a csproj had to be a supported target.
This change allows a project to be converted if it has a supported
target imported anywhere in the csproj. These projects will lose
custom targets in the conversion, but we can still do a pretty good
job of converting in many cases.